requirements.txt
Django
pillow
django-environ
psycopg2-binary
gunicorn
django-heroku
django-storages
boto3